pyproject.toml
[project]
name = "tahtah"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.117.1",
    "ipykernel>=6.30.1",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
]
